Rename logging helper to avoid import conflict
